backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
python-dotenv
openai
python-jose
pydantic
asyncpg
passlib[bcrypt]
python-dotenv
pydantic-settings
alembic
pydantic[email]
python-multipart
authlib
httpx
itsdangerous
pyjwt